// web/src/shared/api/generated/model/getValidationQueueParams.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Elmanhg.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { QuestionDifficulty } from './questionDifficulty';
import type { QuestionType } from './questionType';

export type GetValidationQueueParams = {
unitId?: string;
lessonId?: string;
type?: QuestionType;
difficulty?: QuestionDifficulty;
/**
 * @pattern ^-?(?:0|[1-9]\d*)$
 */
minAgeDays?: number | string;
reviewSessionId?: string;
/**
 * @pattern ^-?(?:0|[1-9]\d*)$
 */
pageNumber?: number | string;
/**
 * @pattern ^-?(?:0|[1-9]\d*)$
 */
pageSize?: number | string;
};