// web/src/shared/api/generated/model/pageDataOfValidationQueueItemResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Elmanhg.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { ValidationQueueItemResult } from './validationQueueItemResult';

export interface PageDataOfValidationQueueItemResult {
  items?: ValidationQueueItemResult[];
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  pageNumber?: number | string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  pageSize?: number | string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  totalItems?: number | string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  totalPages?: number | string;
}